Fix build warning generated by ICC and MSVC compilers.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@6725 6f19259b-4bc3-4df7-8a09-765794883524

+++ OptionRomPkg/CirrusLogic5430Dxe/Edid.c
@@ -118,7 +118,7 @@ ReadEdidData (
   UINT64            Signature;
 
   for (Index = 0; Index < EDID_BLOCK_SIZE * 2; Index ++) {
-    I2cReadByte (Private->PciIo, 0xa0, Index, &EdidData[Index]);
+    I2cReadByte (Private->PciIo, 0xa0, (UINT8)Index, &EdidData[Index]);
   }
 
   //
